+++ src/utils/__tests__/get-code-from-ast.spec.ts
@@ -1,10 +1,7 @@
-import { ParserOptions, parse as babelParser } from '@babel/parser';
 import { format } from 'prettier';
 import { expect, test } from 'vitest';
 
-import { extractASTNodes } from '../extract-ast-nodes';
 import { getCodeFromAst } from '../get-code-from-ast';
-import { getExperimentalParserPlugins } from '../get-experimental-parser-plugins';
 import { getImportNodes } from '../get-import-nodes';
 import { getSortedNodes } from '../get-sorted-nodes';
 
@@ -28,7 +25,7 @@ import a from 'a';
         importOrderSortByLength: null,
         importOrderSideEffects: true,
     });
-    const formatted = getCodeFromAst(sortedNodes, [], code, null);
+    const formatted = getCodeFromAst(sortedNodes, code);
     expect(await format(formatted, { parser: 'babel' })).toEqual(
         `// first comment
 // second comment
@@ -38,32 +35,6 @@ import g from "g";
 import k from "k";
 import t from "t";
 import z from "z";
-`,
-    );
-});
-
-test('it renders directives correctly', async () => {
-    const code = `
-    "use client";
-// first comment
-import b from 'b';
-import a from 'a';`;
-
-    const parserOptions: ParserOptions = {
-        sourceType: 'module',
-        plugins: getExperimentalParserPlugins([]),
-    };
-    const ast = babelParser(code, parserOptions);
-    if (!ast) throw new Error('ast is null');
-    const { directives, importNodes } = extractASTNodes(ast as any);
-
-    const formatted = getCodeFromAst(importNodes, directives, code, null);
-    expect(await format(formatted, { parser: 'babel' })).toEqual(
-        `"use client";
-
-// first comment
-import b from "b";
-import a from "a";
 `,
     );
 });
